+++ apps/meteor/app/threads/client/messageAction/unfollow.ts
@@ -17,6 +17,7 @@ Meteor.startup(() => {
 			id: 'unfollow-message',
 			icon: 'bell-off',
 			label: 'Unfollow_message',
+			type: 'interaction',
 			context: ['message', 'message-mobile', 'threads', 'federated', 'videoconf', 'videoconf-threads'],
 			async action(_, { message }) {
 				if (!message) {