+++ apps/meteor/client/views/room/composer/messageBox/MessageBox.tsx
@@ -81,7 +81,7 @@ type MessageBoxProps = {
 	rid: IRoom['_id'];
 	tmid?: IMessage['_id'];
 	readOnly: boolean;
-	onSend?: (params: { value: string; tshow?: boolean }) => Promise<void>;
+	onSend?: (params: { value: string; tshow?: boolean; previewUrls?: string[] }) => Promise<void>;
 	onJoin?: () => Promise<void>;
 	onResize?: () => void;
 	onTyping?: () => void;
@@ -90,6 +90,7 @@ type MessageBoxProps = {
 	onNavigateToNextMessage?: () => void;
 	onUploadFiles?: (files: readonly File[]) => void;
 	tshow?: IMessage['tshow'];
+	previewUrls?: string[];
 	subscription?: ISubscription;
 	showFormattingTips: boolean;
 	isEmbedded?: boolean;
@@ -107,6 +108,7 @@ const MessageBox = ({
 	onTyping,
 	readOnly,
 	tshow,
+	previewUrls,
 }: MessageBoxProps): ReactElement => {
 	const chat = useChat();
 	const t = useTranslation();
@@ -163,6 +165,7 @@ const MessageBox = ({
 		onSend?.({
 			value: text,
 			tshow,
+			previewUrls,
 		});
 	});